add new ProductCard Component
